ci: fix connect bump-version script
